Imported Upstream version 2.58-svn37702

+++ source/blender/editors/armature/armature_intern.h
@@ -1,5 +1,5 @@
 /*
- * $Id: armature_intern.h 36278 2011-04-21 17:25:58Z campbellbarton $
+ * $Id: armature_intern.h 36644 2011-05-12 16:47:36Z campbellbarton $
  *
  * ***** BEGIN GPL LICENSE BLOCK *****
  *
@@ -206,7 +206,7 @@ void POSE_OT_propagate(struct wmOperatorType *ot);
 /* editarmature.c */
 
 EditBone *make_boneList(struct ListBase *edbo, struct ListBase *bones, struct EditBone *parent, struct Bone *actBone);
-void BIF_sk_selectStroke(struct bContext *C, const short mval[2], short extend);
+void BIF_sk_selectStroke(struct bContext *C, const int mval[2], short extend);
 
 /* duplicate method */
 void preEditBoneDuplicate(struct ListBase *editbones);
